requirements.txt
Django==4.2.7
django-crispy-forms==2.1
coverage==7.6.10
crispy-bootstrap5==2023.10
Pillow==10.1.0
cloudinary==1.42.2
django-cloudinary-storage==0.3.0